Started working on ScoutFragment -- Example code
I started making progress on the Scouting Fragment. As of now it makes an API call that takes a string of comma separated names. I parse the id's from the response and make an API call for each id. I then update the UI onSuccess for each API call. The whole time I keep a count variable to keep track of when to disable the progress bar. I added lots of comments since this is our first time using all of the components together.